Navigational buttons and interactive carousel
Updated each page with navigation buttons to each page. This allows easy navigation between the pages. Updated the carousel to allow it to be interactive, you can now move the carousel forward or backwards with mouse key. I wanted to add a bit more functionality to the carousel like changing pages and auto play.