Updated Pages with info and updated some navigational buttons
Updated the Goals and potential users pages with info from the design document. I also updated the pages with a home button to link back to the start of the app. This allows the user to quickly get back to the beginning if need be. Fixed some navigational buttons not going to the right page,