Adding collapsible list
Updated each page with collapsible lists to reduce how long each page was. It also makes it much easier to read this way. Each section of a page is split up into a list allowing user to read a specific part of that page with ease. Used code examples from jquery website.